Rename default chat agent to basic coding agent
