+++ packages/ai-chat/src/common/general-coding-chat-agent.ts
