+++ packages/ai-chat/src/common/index.ts
@@ -21,5 +21,5 @@ export * from './chat-request-parser';
 export * from './chat-service';
 export * from './chat-variables';
 export * from './command-chat-agents';
-export * from './default-chat-agent';
+export * from './general-coding-chat-agent';
 export * from './delegating-chat-agent';
